Update to Keycloak 26.5.3 and terraform to 1.14.3 and small fixes (#1451)
* Update to Keycloak 26.5.2 and terraform to 1.14.3 and small fixes


* Added fix for missing authorization resource by default


* Update to KC26.5.3


* Removed references to default authorization policy fixed unstable token lifetime tests not using random durations anymore


* Pick better session timings to be compatible with other tests


* Better timeout values for client tests


---------

+++ keycloak/openid_client.go
@@ -152,9 +152,13 @@ func (keycloakClient *KeycloakClient) NewOpenidClient(ctx context.Context, clien
 			if err != nil {
 				return err
 			}
-			err = keycloakClient.DeleteOpenidClientAuthorizationResource(ctx, resource.RealmId, resource.ResourceServerId, resource.Id)
-			if err != nil {
-				return err
+			// Only attempt to delete if the default resource exists
+			// (Keycloak 26.5+ doesn't create default resources automatically)
+			if resource != nil {
+				err = keycloakClient.DeleteOpenidClientAuthorizationResource(ctx, resource.RealmId, resource.ResourceServerId, resource.Id)
+				if err != nil {
+					return err
+				}
 			}
 		}
 	}
